+++ Reconnect/Views/BrowserView.swift
@@ -44,8 +44,6 @@ struct BrowserView: View {
         }
         .toolbar(id: "main") {
 
-            StatusToolbar(applicationModel: applicationModel)
-
             NavigationToolbar(browserModel: browserModel)
 
             ToolsToolbar(browserModel: browserModel)
